requirements.txt
streamlit
opencv-python-headless
numpy
mtcnn
keras-facenet
tensorflow-cpu
pandas